Split tests into separate files for each command and query and enable asserting on log messages
