+++ application/account-management/Tests/Tenants/DeleteTenantTests.cs
@@ -1,0 +1,68 @@
+using System.Net;
+using FluentAssertions;
+using PlatformPlatform.AccountManagement.Database;
+using PlatformPlatform.SharedKernel.Tests;
+using PlatformPlatform.SharedKernel.Tests.Persistence;
+using PlatformPlatform.SharedKernel.Validation;
+using Xunit;
+
+namespace PlatformPlatform.AccountManagement.Tests.Tenants;
+
+public sealed class DeleteTenantTests : EndpointBaseTest<AccountManagementDbContext>
+{
+    [Fact]
+    public async Task DeleteTenant_WhenTenantDoesNotExists_ShouldReturnNotFound()
+    {
+        // Arrange
+        var unknownTenantId = Faker.Subdomain();
+
+        // Act
+        var response = await AuthenticatedHttpClient.DeleteAsync($"/api/account-management/tenants/{unknownTenantId}");
+
+        //Assert
+        await response.ShouldHaveErrorStatusCode(HttpStatusCode.NotFound, $"Tenant with id '{unknownTenantId}' not found.");
+
+        TelemetryEventsCollectorSpy.AreAllEventsDispatched.Should().BeFalse();
+    }
+
+    [Fact]
+    public async Task DeleteTenant_WhenTenantHasUsers_ShouldReturnBadRequest()
+    {
+        // Arrange
+        var existingTenantId = DatabaseSeeder.Tenant1.Id;
+
+        // Act
+        var response = await AuthenticatedHttpClient.DeleteAsync($"/api/account-management/tenants/{existingTenantId}");
+        TelemetryEventsCollectorSpy.Reset();
+
+        // Assert
+        var expectedErrors = new[]
+        {
+            new ErrorDetail("Id", "All users must be deleted before the tenant can be deleted.")
+        };
+        await response.ShouldHaveErrorStatusCode(HttpStatusCode.BadRequest, expectedErrors);
+
+        TelemetryEventsCollectorSpy.AreAllEventsDispatched.Should().BeFalse();
+    }
+
+    [Fact]
+    public async Task DeleteTenant_WhenTenantHasNoUsers_ShouldDeleteTenant()
+    {
+        // Arrange
+        var existingTenantId = DatabaseSeeder.Tenant1.Id;
+        var existingUserId = DatabaseSeeder.User1.Id;
+        await AuthenticatedHttpClient.DeleteAsync($"/api/account-management/users/{existingUserId}");
+        TelemetryEventsCollectorSpy.Reset();
+
+        // Act
+        var response = await AuthenticatedHttpClient.DeleteAsync($"/api/account-management/tenants/{existingTenantId}");
+
+        // Assert
+        response.ShouldHaveEmptyHeaderAndLocationOnSuccess();
+        Connection.RowExists("Tenants", existingTenantId).Should().BeFalse();
+
+        TelemetryEventsCollectorSpy.CollectedEvents.Count.Should().Be(1);
+        TelemetryEventsCollectorSpy.CollectedEvents[0].Name.Should().Be("TenantDeleted");
+        TelemetryEventsCollectorSpy.AreAllEventsDispatched.Should().BeTrue();
+    }
+}
